// src/types/character.types.ts
export interface LocationInfo {
    name: string;
    url: string;
}

export interface Character {
    id: number;
    name: string;
    status: string;
    species: string;
    type: string;
    gender: string;
    image: string;
    url: string;
    created: string;
    origin: LocationInfo;
    location: LocationInfo;
    episode: string[];
}

export interface ApiResponse {
    info: {
        count: number;
        pages: number;
        next: string | null;
        prev: string | null;
    };
    results: Character[];
}